Move components and make core logic directory
